change the service name to for postgres to db
The name of the postgres for dhis has been changed back to db. This had previously been changed to make the name  a bit more descriptive but had to be changed back to db because the dhis container which is dependent on the postgres fails to start up as its always trying to connect to the postgres on host "db". Think this is a issue in the dhis image being used

MENC-18