Add logic to existing scripts to remove config contaiers
These containers were not cleaned up with any previous script. This left
them behind after the rest of the system was removed.
This is an issue for two reasons:
Firstly, the containers were attached to the docker network of the
larger system. This prevents the network from being cleaned up on system
destroy as tehre are containers using it. If the system is restarted the
existing containers won't be able to communicate on the new network as
they are already assigned on the old network. This will throw errors on
config container restart.
Second, these containers do take a bit of space, therefore when they are
no longer needed it would be best to remove them completely.

TRACE-171