Add function for checking if OpenHIM Core is alive
The previous method for ensuring the OpenHIM core API was running wasnt very consistent.
If a computer is running slow for various reason, the 20 second sleep period might not be enough.
This caused the mediator services to executing assuming the API is up and running but the API was still busy starting.
The services would therefor fail to connect and register to the OpenHIM