Add unused label field lookup filter query param
JSON does not allow comments in the code.
The value in the lookup query param filter is uuid referencing a DHIS2
field. This is not obvious and users can easily overlook this necessary
configuration point.
The label field explains what is happening and won't affect the
functionality of the endpoint

TRACE-156